requirements/development.txt
pytest
pytest-django
pytest-mock
flake8
